添加异步流程
Bump aws-sdk-s3 from 1.184.0 to 1.185.0

Bumps [aws-sdk-s3](https://github.com/aws/aws-sdk-ruby) from 1.184.0 to 1.185.0.
- [Release notes](https://github.com/aws/aws-sdk-ruby/releases)
- [Changelog](https://github.com/aws/aws-sdk-ruby/blob/version-3/gems/aws-sdk-s3/CHANGELOG.md)
- [Commits](https://github.com/aws/aws-sdk-ruby/commits)

---
updated-dependencies:
- dependency-name: aws-sdk-s3
  dependency-version: 1.185.0
  dependency-type: direct:production
  update-type: version-update:semver-minor
...


Bump rspec-rails from 7.1.1 to 8.0.0

Bumps [rspec-rails](https://github.com/rspec/rspec-rails) from 7.1.1 to 8.0.0.
- [Changelog](https://github.com/rspec/rspec-rails/blob/main/Changelog.md)
- [Commits](https://github.com/rspec/rspec-rails/compare/v7.1.1...v8.0.0)

---
updated-dependencies:
- dependency-name: rspec-rails
  dependency-version: 8.0.0
  dependency-type: direct:development
  update-type: version-update:semver-major
...

+++ db/schema.rb
@@ -77,7 +77,7 @@ ActiveRecord::Schema[8.0].define(version: 2025_05_04_111514) do
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.string "api_identifier"
-    t.integer "app_entity_id", null: false
+    t.bigint "app_entity_id", null: false
     t.string "webhook_url"
     t.index ["api_identifier"], name: "index_dynamic_tables_on_api_identifier", unique: true
   end
